Getting the JSON by making HTTP call, I am adding a Async class GetExchangeData to make http calls on background thread. Add the following method in main activity class.
In onPreExecute() method progress dialog is shown before making the http call.

In doInBackground() method, makeServiceCall() is called to get the json from url. Once the json is fetched, it is parsed and each contact is added to array list.

In onPostExecute() method the progress dialog is dismissed and the array list data is displayed in list view using an adapter.

Also note that I have used getJSONObject() method depending on the type of node.